Update app.py
We added DuckDuckGo search results alongside Google search results and the Google fat checker to make sure this is more accurate.

My apologies go to Robert Skobel and the community for having soft launched this without more than one source to verify news which caused a false positive report on:

Alexandre de Moraes blocks financial accounts of Starlink in Brazil.

Which was true. We apologize and hope this gets better.